Added customers growing and shrinking to despawn and respawn
Customers now grow and shrink to disappear and reappear. If they get interrupted by the ticket turning in before their fully grown or ticket being made, they just jump to do the start of growing or shrinking when appropriate. Size of customers do not matter, but adjust the parent when changing size.