fix: Fixed material for when tomato block is sliced
When slicing a food block, no material exists on the inside of the object.

To not have sliced objects appear empty on some sides, or fully invisible, the sliced objects use a matte material color to fill their missing regions.

Somehow, the tomato red material was changed to use the tomato's base map, which shouldn't be the case.

Fixed to use simple red and tested in enhancedCutting scene.