feat: Implemented combine sound effect
When combining objects to the plate, you now hear a "Pop" sound effect.

Also, modified CollisionDetector script to not play its sound if the collided object is a food object as this would play two sounds for the same collision (at least for how we have food objects setup with two colliders)

Added in steam hiss sound effects to hopefully help mask audio cutoff when an cookable object transitions to its next state.